+++ src/main/java/core/model/player/Player.java
@@ -12,10 +12,7 @@ import core.model.match.Weather;
 import core.net.OnlineWorker;
 import core.rating.RatingPredictionManager;
 import core.training.*;
-import core.util.HODateTime;
-import core.util.HOLogger;
-import core.util.Helper;
-import core.util.HelperWrapper;
+import core.util.*;
 import org.jetbrains.annotations.Nullable;
 
 import java.time.Duration;
@@ -332,7 +329,8 @@ public class Player extends AbstractTable.Storable {
     private String motherclubName;
     private Integer matchesCurrentTeam;
     private int hrf_id;
-
+    private Integer htms = null;
+    private Integer htms28 = null;
     public int getGameStartingTime() {
         return GameStartingTime;
     }
@@ -340,6 +338,7 @@ public class Player extends AbstractTable.Storable {
     public void setGameStartingTime(int gameStartingTime) {
         GameStartingTime = gameStartingTime;
     }
+
 
     //~ Constructors -------------------------------------------------------------------------------
 
@@ -1714,6 +1713,34 @@ public class Player extends AbstractTable.Storable {
 
     public void setLastMatchId(Integer v) {
         this.m_lastMatchId = v;
+    }
+
+    public Integer getHtms28() {
+        if ( htms28 == null){
+            htms28 = Htms.htms28(getSkills(), this.m_iAlter, this.m_iAgeDays);
+        }
+        return htms28;
+    }
+
+    public Integer getHtms(){
+        if (htms == null){
+            htms= Htms.htms(getSkills());
+        }
+        return htms;
+    }
+
+    private Map<Integer, Integer> getSkills() {
+
+        var ret = new HashMap<Integer, Integer>();
+        ret.put(KEEPER, getGKskill());
+        ret.put(DEFENDING, getDEFskill());
+        ret.put(PLAYMAKING, getPMskill());
+        ret.put(WINGER, getWIskill());
+        ret.put(PASSING, getPSskill());
+        ret.put(SCORING, getSCskill());
+        ret.put(SET_PIECES, getSPskill());
+
+        return ret;
     }
 
     public static class Notes extends  AbstractTable.Storable{