+++ src/main/java/core/training/type/DefensivePositionsWeeklyTraining.java
@@ -1,14 +1,9 @@
 package core.training.type;
 
-import java.util.List;
-
 import core.constants.TrainingType;
 import core.constants.player.PlayerSkill;
-import core.model.StaffMember;
-import core.model.UserParameter;
 import core.model.player.IMatchRoleID;
 import core.model.player.MatchRoleID;
-import core.model.player.Player;
 import core.training.WeeklyTrainingType;
 
 public class DefensivePositionsWeeklyTraining extends WeeklyTrainingType {
@@ -39,7 +34,7 @@ public class DefensivePositionsWeeklyTraining extends WeeklyTrainingType {
 		_PrimaryTrainingSkillOsmosisTrainingPositions = new int[] { IMatchRoleID.leftForward,
 				IMatchRoleID.centralForward, IMatchRoleID.rightForward};
 		_PrimaryTrainingBaseLength = DefendingWeeklyTraining.instance().getBaseTrainingLength(); // old was (float) 7.2
-		_PrimaryTrainingSkillBaseLength = (_PrimaryTrainingBaseLength + UserParameter.instance().TRAINING_OFFSET_DEFENDING) * 2;
+//		_PrimaryTrainingSkillBaseLength = (_PrimaryTrainingBaseLength + UserParameter.instance().TRAINING_OFFSET_DEFENDING) * 2;
 	}
 	public static WeeklyTrainingType instance() {
         if (m_ciInstance == null) {