+++ src/main/java/core/training/type/ShortPassesWeeklyTraining.java
@@ -1,14 +1,9 @@
 package core.training.type;
 
-import java.util.List;
-
 import core.constants.TrainingType;
 import core.constants.player.PlayerSkill;
-import core.model.StaffMember;
-import core.model.UserParameter;
 import core.model.player.IMatchRoleID;
 import core.model.player.MatchRoleID;
-import core.model.player.Player;
 import core.training.WeeklyTrainingType;
 
 public class ShortPassesWeeklyTraining extends WeeklyTrainingType {
@@ -38,7 +33,7 @@ public class ShortPassesWeeklyTraining extends WeeklyTrainingType {
 				IMatchRoleID.leftBack, IMatchRoleID.rightBack, IMatchRoleID.leftCentralDefender,
 				IMatchRoleID.middleCentralDefender, IMatchRoleID.rightCentralDefender };
 		_PrimaryTrainingBaseLength = (float) 4.2989; // old was 2.8
-		_PrimaryTrainingSkillBaseLength = _PrimaryTrainingBaseLength + UserParameter.instance().TRAINING_OFFSET_PASSING;
+//		_PrimaryTrainingSkillBaseLength = _PrimaryTrainingBaseLength + UserParameter.instance().TRAINING_OFFSET_PASSING;
 	}
 	
 	public static WeeklyTrainingType instance() {